Move definition of columns to separate file as advised in pull request.
