--- src/sandbox/mirek/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const MirekSandboxPage: React.FC = () => {
-  return (
-    <>
-      <h1>Mirek Sandbox</h1>
-    </>
-  );
-};
-
-export default MirekSandboxPage;
