--- src/sandbox/wiktor/index.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const WiktorSandboxPage: React.FC = () => {
-  return (
-    <>
-      <h1>Wiktor Sandbox</h1>
-    </>
-  );
-};
-
-export default WiktorSandboxPage;
